google tag manager with consent
removed google analytics, and replace it with GTM with consent.
It will make it easier to deals with ads (if I end up adding ads),
some google services have built-in consent, so it's easier to deal with cookies.
I've read that when consent is denied, GA4 is still able to collect some data
without cookies, but my tests didn't show anything, you can see events in GTM
but nothing is receive by GA4, so... whatever, I need to move.